[#27387] YSQL: Improve logging for incremental catalog cache refresh
Summary:
This diff makes changes to improve logging incremental catalog cache refresh:

1. Added a reason when a full catalog refresh is needed
Note that sometimes we do not need to make a local request to
`TabletServer::GetTserverCatalogMessageLists`. For example, if the PG backend is
lagging far behind that the number of requested catalog versions already exceeds
the maximum queue length, then even if a call to
`TabletServer::GetTserverCatalogMessageLists` is made, we'll get a negative
answer and then a full refresh is performed. If the PG backend's local catalog
version is 0, that means a cache reset is requested by PG. Even if we make a
call to `TabletServer::GetTserverCatalogMessageLists`, we'll get a negative
answer because catalog version 1 does not have any invalidation messages.

2. Changed a pointer logging to be more meaningful.

3. Also made a change to skip some metrics accounting when
`yb_is_calling_internal_sql_for_ddl` is true so that we do not have internally
executed SQL statements to interfere normal metrics results.
Jira: DB-16921

Test Plan: ./yb_build.sh --cxx-test pg_catalog_version-test

Reviewers: kfranz, sanketh, mihnea, #db-approvers

Reviewed By: kfranz, #db-approvers

Subscribers: jason, yql

Differential Revision: https://phorge.dev.yugabyte.com/D44312