[PLAT-17290][xCluster] table status metric is cleared when a participating universe has a task in progress
Summary:
Previously, YBA suppressed reporting the xCluster bad status table metric if a task was running on either the source or target universe. This was done to avoid running the RPC calls needed for xCluster metrics in parallel with other RPCs.

However, the DB team has confirmed that these RPCs are read-only and can safely run concurrently with others. This diff removes the restriction, ensuring the metric is always reported.

Test Plan: - Ran a task on the universe and made sure the status continues to be reported

Reviewers: amindrov, spothuraju, amalyshev, daniel

Reviewed By: amalyshev

Differential Revision: https://phorge.dev.yugabyte.com/D44338