[#27248][#2605] CDC: Send single BEGIN and COMMIT records for colocated tables
Summary:
An UPDATE which does not affect any rows generates a WAL op with no write pairs. When a WAL op is encountered by cdcsdk_producer, it first adds a BEGIN record, and then tries to add records based on the write pairs and in the end adds a COMMIT record. For a WAL op with no write pairs, it removes the BEGIN record at the end instead of adding an empty BEGIN-COMMIT in the response.

Previously the [[ https://github.com/yugabyte/yugabyte-db/commit/d19c3664f738e7f7b0d1c3922145746350f55296 | commit ]] tried to fix a similar issue, where when no write pairs were found for the very first WAL op, we added a DDL record and the logic to remove the BEGIN record ended up removing only the DDL record and sending a BEGIN without a corresponding COMMIT record. This revision did not take into consideration the case when a WAL op with no write pairs is received for a colocated tablet (with multiple tables) on subsequent GetChanges calls.

When such a WAL op is encountered the logic for removing the last BEGIN record fails. In such a scenario we add as many BEGIN / COMMIT records as the number of tables residing on the colocated tablet. When a WAL op with no write pairs is encountered for such a tablet, we only remove one BEGIN added to the response. This results in cdcsdk_producer sending BEGIN records without corresponding COMMITs. These BEGINS can cause issues in virtual wal / downstream applications (in case of gRPC model).

This revision fixes the bug by sending a single BEGIN and a single COMMIT record for colocated tables as well. The BEGIN / COMMIT records will no longer contain the table name.

The multiple BEGIN / COMMIT records for colocated tables are anyway not required anymore. Virtual WAL filters out all the BEGIN / COMMIT records except one (the first BEGIN and the last COMMIT are kept). Earlier this was required with gRPC model, when transactional ordering was enabled. This mode is now deprecated and anyone requiring transactional ordering is recommended to use logical replication.

The `table_name` field in BEGIN / COMMIT records is also not used anywhere in logical replication and hence can be removed without any repercussions. For gRPC connector a minor [[ https://github.com/yugabyte/debezium-connector-yugabytedb/pull/372 | change ]] is needed to accommodate the removal of `table_name` field in transactional records
Jira: DB-16731, DB-2038

Test Plan:
New tests:

  - ./yb_build.sh --cxx-test integration-tests_cdcsdk_consumption_consistent_changes-test --gtest_filter CDCSDKConsumptionConsistentChangesTest.TestColocatedSingleShardUpdateAffectingNoRows
  - ./yb_build.sh --cxx-test integration-tests_cdcsdk_consumption_consistent_changes-test --gtest_filter CDCSDKConsumptionConsistentChangesTest.TestColocatedMultiShardUpdateAffectingNoRows

Ran the grpc connector unit tests and the `TestPgReplicationSlot` tests to confirm that removing the tableName from the BEGIN and COMMIT message and also having a single BEGIN/COMMIT pair while dealing with colocated tables does not break anything on the client side.

Reviewers: skumar, asrinivasan, vkushwaha

Reviewed By: asrinivasan

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D44144