[#27392] Docdb: Fix YSQL per session docdb traces interaction with pg_client_use_shared_memory
Summary:
Currently, Per txn tracing based on `yb_enable_docdb_tracing` does not work when `pg_client_use_shared_memory` is enabled.
Given that `pg_client_use_shared_memory` is on by default, this feature has been unusable.

The diff fixes this issue, by ensuring that, if there is no transactional context, or rpc context,
a trace is locally created and printed at the end of DoPerform.

If the current operation is done in the context of a transaction, a child trace will be added
to the transaction's trace; and the whole transaction trace will be printed at the end of the
transaction.

If there is no transaction set, then the trace will be printed at the end of the current
Perform rpc/DoPerform operation.
Jira: DB-16926

Test Plan: yb_build.sh release --cxx-test pg_mini-test --gtest_filter PgMiniTestTracing/PgMiniTestTracing.Tracing/*

Reviewers: rthallam, bkolagani

Reviewed By: bkolagani

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D44358