[PLAT-17164] Make all YBClient invocations use try-with-resources
Summary:
Right now a lot of placse, where we use YBClient, don't do it via try-with-resources.
This makes it hard to make sure we close the client always.
Making a change to use try-with-resources always.
Was not able to find a place where the client would not be closed, though - so this is more a refactoring.

Test Plan: itests

Reviewers: vbansal, skurapati

Reviewed By: skurapati

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D44365