[BACKPORT 2025.1][#27064] Docdb: Preserve next column ID of DocDB Table at import_snapshot
Summary:
Original commit: 9912fdc077dd49ed1dc6be0c631209affb6454bf / D44280
Clean backport. No merge conflicts.
Prior to this diff, the next column ID for a DocDB table on the restore side was calculated as `schema.max_col_id() + 1` where `max_col_id` is the maximum column id that exists in the schema of the DocDB table at the backup side.
This means that the next column id on the backup and restore sides might not always be the same. Consider the following scenario:
 - `CREATE TABLE t(x INT,y INT);` where [column_name:column_id] mappings are: [x:1],[y:2] and `next_column_id=3`
 - `ALTER TABLE t DROP COLUMN y;`
 - Take a backup at this point. `[x:1], max_col_id = 1, next_column_id = 3`
 - Restore the previous backup which creates the table t with schema [x:1] and `next_column_id=max_col_id+1=2`.
We can see that the Backup and restore sides have a different `next_column_id` due to the dropped column y on the backup side. This causes issues for xClsuter DDL replication. Additionally, it can cause errors on the restore side. For example, if after restore we performed:
 - `ALTER TABLE t ADD COLUMN z TEXT;`
 This adds a new column [z:2], which reuses the ID of the previously dropped column y. In case the deleted column values haven't been garbage collected, the tablet now has conflicting data for column ID 2. If the two columns (the dropped ones and the newly created ones) are of different types, then we get the following error when reading data from the table:
```
ERROR:  Unexpected value type: kInt32
```
However, if the two columns are of the same type, then we might return incorrect data to the end user.

The diff fixes the issue by setting the restored table’s `next_column_id` to match the `schema.next_column_id` from the corresponding DocDB table schema from the backup side. This ensures column ID continuity and avoids collisions with previously dropped columns.
Jira: DB-16545

Test Plan:
./yb_build.sh release --cxx-test yb-backup-cross-feature-test --gtest-filter YBBackupTest.TestPreservingNextColumnId

./yb_build.sh release --cxx-test xcluster_ddl_replication-test --gtest_filter XClusterDDLReplicationTest.DocdbNextColumnAboveLastUsedColumn

Reviewers: mlillibridge, asrivastava

Reviewed By: asrivastava

Subscribers: slingam, ybase, hsunder

Differential Revision: https://phorge.dev.yugabyte.com/D44374