[BACKPORT 2025.1][#27297] YSQL: add multirange catalog entries
Summary:
Add catalog entries for the lsm access method for multirange types (see
upstream postgres commit 6df7a9698bb036610c1e8c6d375e1be38cb26d5f).
This mainly duplicates catalog entries for btree+multirange to
lsm+multirange.  This is similar to the duplication that has already
been done for btree+range to lsm+range.  There is a catalog pg_range
that has only btree+multirange, and that is left alone.

With these entries, creating an lsm index on multirange column is one
step closer to being supported.  Before, it fails due to lack of catalog
entries:

    ERROR:  data type nummultirange has no default operator class for access method "lsm"
    HINT:  You must specify an operator class for the index or define a default operator class for the data type.

Now, it fails due to YB not supporting the type as a key (see
yb_type.c YbTypeEntityTable):

    ERROR:  INDEX on column of type 'NUMMULTIRANGE' not yet supported

Port the remaining queries in regress test multirangetypes.  Index
creation fails, but keep the queries related to that even if they don't
exercise the index since, if those are removed, there aren't many
queries left to test.

Add a new YSQL migration upgrade snapshot for testing because adding a
migration script is currently not possible with this kind of catalog
change.  This should be fine as long as this change goes in to 2025.1.0.
Upgrading on the 2.25.x line may miss these catalog entries, but that
isn't the end of the world because these entries aren't on the critical
path.

Backport note:
[[ https://github.com/yugabyte/yugabyte-db-ysql-catalog-snapshots/releases/tag/v2025.1.0.0-pg15-12-4 | A new snapshot for 2025.1 ]] has been uploaded to the snapshots repo; this diff updates the tests on 2025.1 to use this snapshot.
This is the first time the base snapshot is taken from a dot version (in this case, v78.2). Fix `TestYsqlUpgrade#migratingIsEquivalentToReinitdb` to take into account the minor version when calculating the expected number of migrations applied.

Jira: DB-16793

Test Plan:
On Linux:
```
./yb_build.sh release --java-test org.yb.pgsql.TestYsqlUpgrade#migratingIsEquivalentToReinitdb
./yb_build.sh asan --java-test org.yb.pgsql.TestYsqlUpgrade#migratingIsEquivalentToReinitdb
```

On Mac:
```
./yb_build.sh release --java-test org.yb.pgsql.TestYsqlUpgrade#migratingIsEquivalentToReinitdb
```

On Almalinux 8:

    ./yb_build.sh fastdebug --gcc11 --java-test TestPgRegressPgTypesUDT

Close: #27297
Original commit: 99a7de4e8049de467ef5c5bcb8b2f2c754ef9dc7 / D42128

Reviewers: jason, #db-approvers

Reviewed By: jason, #db-approvers

Subscribers: svc_phabricator, yql

Differential Revision: https://phorge.dev.yugabyte.com/D44349