[#27426] docdb: Post split compaction may end up too early not compacting some parent files
Summary:
Post split compaction may stop too early, skipping some SST files that should be compacted, if
a background compaction is triggered while the post-split compaction is still in progress. This
issue can occur only when the post-split compaction operates in multiple iterations, and
a background compaction starts between those iterations. As a reminder: multiple iterations are
used to avoid locking too many files at once by limiting the input size per compaction iteration.
The total size of input files for one iteration should not exceed
`post_split_compaction_input_size_threshold_bytes` (default: 256 MB).

The problem comes from an incorrect assumption in the post-split compaction logic:
if the compaction picker encounters an SST file marked as `being compacted`, it stops and returns
the current set of selected files. If the returned set is empty, the post-split compaction is
considered complete. This logic breaks if a background compaction is triggered just before
the post-split compaction picker attempts to fetch the next subset of SST files and if
the background compaction picks the subset of most earliest files: the very first file becomes
marked as `being compacted`, which lead to the picker stops immediately and returns an empty set,
causing the post-split compaction to terminate prematurely.

Despite the fact the issue is a generic and may be observed in any configuration where tablet
splitting happens, the easiest way to capture the issue is to configure default table TTL and
turn on TTL file expiration feature in order to limit background compactions by SST file size.

Example:
```
1. Default table TTL is confugred, TTL file expiration is configured, rocksdb max file size
   for compaction is set to 3 GB, post split compaction input threshold is 256 MB.

2. SST files before post split compaction:
   SST1 [seq1], SST2 [seq2], SST3 [seq3], SST4 [seq4], SST5 [seq5], SST6 [seq6], SST7 [seq7]
   SST1 ... SST7 are 3.1 GB each

3. The state after 5 out of 6 iterations (where only 1 SST is taken by during each iteration):
   SST8 [seq1], SST9 [seq2], SST10 [seq3], SST11 [seq4], SST12 [seq5], SST6 [seq6], SST7 [seq7]
   SST8 ... SST12 are 1.5 GB each, SST6 and SST7 are 3.1 GB

4. Background compaction kicks in immediately after SST11 has been flushed.

5. Background compaction takes SST8...SST12 by size amp criteria and markis them `being compacted`.

6. Post split compaction runs the next iteration but it sees the first file by sequence, SST8, is
   locked and immediately stops compacing due to no files got picked on this iteration.

7. Background compaction processes SST8...SST12 producing SST13 [seq1] of size 7.5 GB (5 x 1.5 GB).

8. The final set of SST files after both post split and background compactions are completed:
   SST13 [seq1, 7.5 GB], SST6 [seq6, 3.1 GB], SST7 [seq7, 3.1 GB]
```

The fix: if post split compaction picker meets the locked file (under compaction), do compact files
which are currently collected, or move to the next file (if nothing got picked yet).

This fix is required to be backported into all the active releases, but we don't necessary need
to backport the unit test for the fix. Hence the unit test is being landed by a separated diff
https://phorge.dev.yugabyte.com/D44537
Jira: DB-16966

Test Plan: Jenkins

Reviewers: timur, rthallam

Reviewed By: timur, rthallam

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D44394