[PLAT-17758] Implement basic integration tests for yba-installer
Summary:
This creates a framework called `containertest` that can be leveraged
to start containers during a go test, copy files to the container,
and run commands inside the container. It is set up to also be able to
pass through systemd, to allow yba installer to manage services

NOTE: Currently, this will not work for user level systemd.

To validate, I also implemented a couple of cert tests

Test Plan: ran the integration tests

Reviewers: muthu, nsingh

Reviewed By: muthu

Differential Revision: https://phorge.dev.yugabyte.com/D44495