[BACKPORT 2025.1][#27380]xCluster: Fix transactional writes on xCluster range partitioned tables
Summary:
For transactional writes in YugabyteDB, we generate a sequence of intents records followed by a single APPLY record(one per transaction) in the WAL. This happens on every tablet that was written to by the transaction. Normally the APPLY would move all intents associated with the transaction to the regular db. xCluster replicates these records to the target as is, so it also gets intents and APPLY. When we have 2 producer tablets that get replicated to 1 consumer tablet we will get 2 sequence of intents and 2 APPLY op(one for each producer tablet). These are ordered for each producer tablet, but between 2 (or more) tablets these can get interleaved.
Ex:
```
Tablet -> T
Intent -> I
T1 I1, T1 I2, T2 I3, T1 APPLY, T2 I4, T2 APPLY
```

If we process this as is the `T1 APPLY` would move `T2 I3` to regular db incorrectly. Diff c543bd0d12cf54957cad3cb0034df5cbcbc3cd06/D30388 attempted to fix this situtation. The fixed involved attaching the key range to each APPLY and using it to only APPLY T1s intents when we process T1s APPLY.
This however does not work for range partitioned tables, since the partition key for hash is encoded whereas for range they are not.

This change fixes this issue by propagating the encoded key range for all partition types, and using that to determine which intents to apply. This logic aligns the xCluster code with other existing components like tablet splitting and compactions.

Also moving `ExternalTxnApplyStateData` to `rocksdb_writer.cc` since its only used in this file.

**Upgrade/Rollback safety**
There are a max of 3 nodes involved for each record replicated with xCluster. The Producer tablet leader, the target poller and the destination tablet leader.
In the case of Bi-directional xCluster both clusters can be upgraded in parallel.
`encoded_start_key` and `encoded_end_key` are added to `CDCRecordPB` thats part of the `GetChangesResponsePB`.
The producer node on the new code unconditionally populates these along with the existing `partition` field. The `partition` field is used in to detect which tablets to route the request to, so it cannot be removed.
The target poller sets `filter_range_encoded` when `filter_start_key` and `filter_end_key` of the `ApplyExternalTransactionPB` are populated with the encoded keys. This is guarded by AutoFlag `xcluster_use_encoded_key_filter`, so the until the upgrade completes (target tserver upgrades to new version) the old logic is used.
The destination tablet leader uses `filter_range_encoded` to determine if it should use the new or old logic.
Jira: DB-16895

Original commit: e4982612817289377ce8a23875ef42ad7c43bb22 / D44302

Test Plan:
non_transactional_batch_writer-test
TEST_P(XClusterYsqlSimpleReplicationTest, Validate) {
TEST_F(XClusterDBScopedTest, RangedPartitionsWithIndex) {
TEST_F(XClusterDBScopedTest, ColocatedRangedPartitionsWithIndex) {

Jenkins: Urgent

Reviewers: jhe, sergei, slingam

Reviewed By: jhe

Subscribers: ybase, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D44423