[#27425] Docdb: Allow export snapshot with tables undergoing DDL verification
Summary:
Currently, `export snapshot` fails in case one of the tables included in the snapshot is in the middle of a DDL operation. This protects against successfully taking a backup while a long-running DDL is taking place during backup creation. In such scenarios, the catalog version check performed at the beginning and end of the current backup workflow might pass as the long-running DDL might not have bumped the catalog version yet. Thus, we rely on failing `export snapshot` to signal the failure of backup creation.

However, with the new backup workflow, "where we support taking backups during DDLs"; We want to allow `export snapshot` even if some of the tables that are part of the snapshot are in the middle of a DDL. The idea is that it is ok to export these tables that are undergoing a DDL verification. On the restore side, we rely on import snapshot to correctly choose the committed DocDB tables from the snapshot and ignore the uncommitted ones.

Therefore, we want to add the ability to allow exporting a snapshot even if some of its tables are undergoing a DDL verification step. We also want to conserve the old behavior, as the old workflow will still be supported by YBDB.

Therefore, a new field `include_ddl_in_progress_tables` is added to the `ListSnapshotsRequestPB`, which controls whether `export snapshot` allows exporting tables that are undergoing DDL verification or not.

The field will be set to true once YBC starts using the new backup workflow, and it is set to false by default. Setting this flag to true also exports the snapshot with format version = 3 instead of 2. Which indicates that this backup must be restored using the relfilenode id to build the table mappings rather than table names.

**Upgrade/Downgrade safety**
A new proto field `include_ddl_in_progress_tables` is added to `ListSnapshotsRequestPB` and is false by default. This means that the default behavior is the old behavior of failing list snapshot in case one of the tables is undergoing a DDL verification.

The old behavior will happen in one of the following cases:
- If the client process runs the old code and doesn't know about the option `include_ddl_in_progress_tables` in `ListSnapshotsRequestPB` while the master process knows about `include_ddl_in_progress_tables`. Then the old behavior will happen which is expected.
- If the client process is running the new code and sets the field to true, but the master is running old code that doesn't know about this field, then `ListSnapshot` will also revert to the old behavior. We are ok with failing the `ListSnapshot` operation until the master is upgraded to the new code, which understands the `include_ddl_in_progress_tables` option. In that case, the client can retry the `ListSnapshot` RPC after the DDL has finished.
Jira: DB-16965

Test Plan:
./yb_build.sh release --cxx-test yb-admin_client-test --gtest-filter ClusterAdminClientTest.ExportSnapshotWithDdlInProgressTables
./yb_build.sh release --cxx-test yb-admin_client-test --gtest-filter ClusterAdminClientTest.ExportSnapshotHasCorrectFormatVersion

Reviewers: sanketh, vkumar, asrivastava

Reviewed By: asrivastava

Subscribers: ybase, slingam

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D44460