[#27460, #27479] DocDb: Table locks. Fix lease establishment after network partition.
Summary:
As reported during testing: after a network partition the master_leader_poller is unable to discover the new master-leader to establish the lease.

This is because ResetProxy only gets called when
 consecutive_failures_ == FLAGS_heartbeat_max_failures_before_backoff
Thus, if a failure were to persist longer; and then a new master leader is established, the poller
is unable to discover the new master.

The fix is to ensure that we ResetProxy whenever
 consecutive_failures_ >= FLAGS_heartbeat_max_failures_before_backoff
so that the new master leader will be discovered.
Jira: DB-16998, DB-17022

Test Plan: yb_build.sh fastdebug --cxx-test 'TEST_F(MultiMasterObjectLockTest, TServerCanAcquireLeaseAfterProlongedPartition) {'

Reviewers: bkolagani, rthallam

Reviewed By: bkolagani

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D44487