[#27488] docdb: Increase cluster balancer remove and over-replication limits
Summary:
As part of increasing the speed of cluster operations, we must increase `load_balancer_max_concurrent_removals` and `load_balancer_max_over_replicated_tablets` (among other flags, which will be adjusted in the future).

**Default flag value changes**
- `load_balancer_max_over_replicated_tablets` controls the maximum number of tablets **per table** that can be over-replicated. **This diff changes the default from 1 to 50.**
  - If we are over this limit for a given table, we will not add any more tablets. This includes tablets that we have started in the current run of the cluster balancer, so it practically limits the number of new moves per run as well.
  - This may cause additional space amplification in the case where removes are slow, which is why we increase that flag in this diff as well
  - Some history: There is no description as to why the flag was added in the original commit (d14d2f3 / D3903), but it was possibly to prevent unbounded space amplification if `RemoveReplica` requests are failing.
  - There have always been guards to prevent adding a replica to a tablet that is **already** over-replicated (see commit bcce2c2 / D153), so the over-replication per tablet is limited to 1.
- `load_balancer_max_concurrent_removals` controls the number of replicas that can be removed per cluster balancer run. **This diff changes the default from 1 to 50**
  - Since removing a replica is a relatively cheap operation, we don't expect this to overwhelm any tserver

**Inefficient behavior when `load_balancer_max_over_replicated_tablets` > 1 fix**
This diff also fixes inefficient behavior in the case where `load_balancer_max_over_replicated_tablets` is set to a value greater than 1. The cluster balancer will move tablets between tservers if the tablet count difference is at least 2. However, this does not account for the fact that we might be removing replicas from the source tserver in the future. For example, if tserver A has 50 tablets and tserver B (in the same AZ) has 0 and `load_balancer_max_over_replicated_tablets=50`, we will add 49 tablets to B, and in subsequent runs remove 25 tablets from A and 24 from B. We should instead just move 25 tablets from A to B.

This diff addresses this by not counting over-replicated replicas as load on the **source** tserver, under the conservative assumption that we might remove them all in a future run. In the worst case, all of these replicas will be removed from **other** tservers, in which case this behavior will cause us to revert to the same behavior we had before this diff (with `load_balancer_max_over_replicated_tablets=1`) because we will wait for there to be no over-replication before we increase trigger new adds / moves.
Jira: DB-17036

Test Plan:
Added the following test to test the over-replication in source load:
`./yb_build.sh --cxx-test master_load_balancer_mocked-test --gtest_filter LoadBalancerMockedTest.ExcludeOverreplicationInSourceLoad`

Also triggered some perf service pipelines for scaling in and out with sysbench / tpcc. Perf studio runs can't be automatically compared so I've added the scaling time results below. I also checked the latencies and they seem in line with the current values.
| Test type | Test id | Result change (old -> new) |
| Scale out with TPCC Single-AZ | 12246102 | Load balancing time: 400s -> 180s, Total scaling time: 1100s -> 820s
| Scale in with Sysbench - Single AZ | 12215102 | Scaling time: 5489s -> 4075s
| Scale in with TPCC - Single AZ | 12215002 | Scaling in time: 652s -> 467.0s
| Scale out with Sysbench - Single AZ | 12214802 | Load balancing time: 4691s -> 4269s
| Scale in with Sysbench - Multi AZ | 12214702 | Scaling time: 5439s -> 4046s
{F371869}

{F371868}

{F371867}

{F373081}

{F371865}

Reviewers: jhe, zdrudi

Reviewed By: zdrudi

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D44501