[PLAT-17733]: Integrate Performance Advisor latest changes into YBA
Summary:
1st commit
1. Rename all file names as Troubleshoot Platform project has changed name to Perf Advisor
2. Rename translation strings
3. Remove TP package and install new perf-advisor package
4. Change data-testId and variable names, api function names

2nd commit
1. Add tabName as query Param when user is in metrics view
2. Resize graph when it shows empty data
3. remove side panel to add and remove metrics (only needed for YBM)

3rd commit
1. Remove console statements
2. Unmount component and remove query param when user moves to a different tab
3. Ensure default tab is shown when on metrics view

Most of the files are just name changes.
Only new file added is PerfAdvisorTabs.tsx which ensures YBA is registered to PA, if yes, route it to PA components
By default this feature will be disabled, this is just for internal use where we want to integrate the new Perf Advisor UI into YBA and YBM based on a runtime config.

Test Plan:
{F361394}

{F361395}

{F361396}

{F361397}

{F361398}

{F361399}

{F361400}

{F361401}

Reviewers: jmak

Reviewed By: jmak

Differential Revision: https://phorge.dev.yugabyte.com/D44311