[BACKPORT 2025.1][PLAT-17266] Add an alert on YBA for the kms key on a universe
Summary:
This diff adds a new metric `ybp_universe_kms_key_status` which is populated by YBA every 1 hour from the Refresh KMS Periodic Background Service.

Logic to populate the metric:
We go through all existing universes for all customers and see if they have EAR enabled. If EAR is enabled, we get the last active KMS config used on it and try to do 2 things:
1. Check if the currently active universe key can be decrypted by the EAR service or not.
2. Do a refresh of the entire KMS config, which basically checks the permissions of the master key, ensures it is available to YBA, does a sample encrypt + decrypt operation on a random test string, updates the TTLs wherever applicable (currently only for Hashicorp vault).

If both the above conditions succeed, it means the KMS key is valid and the metric gets populated to a value of `1`.
If it throws an error or fails any validation, the metric becomes `0`.

Based on that hourly metric, we add a new default alert which alerts the user if the KMS key is invalid or not.

Sample alert:
{F360632}

Original commit: 1a235c49085fb9a90e0a6bb03f68e3681819a052 / D44404

Test Plan:
Manually tested couple of EAR flows on a universe:
1. Enable EAR, ensure key is valid, check no alert.
2. Remove some required key permissions on the KMS, check that the alert is raised within 1-2 minutes. (Also changed the RefreshKmsService to run every 1 min instead of 1 hour for local testing)
3. Add the permissions back, check that the alert is resolved automatically.
4. Deleted the master key on Ciphertrust KMS and recreated a new one with the same name, ensured the alert is still triggered (since the active universe key cannot be decrypted with this new one).

Tested this for AWS and Ciphertrust KMS, rest of the KMS providers work similarly.

Also run itests.
And run UTs.

Reviewers: amalyshev

Reviewed By: amalyshev

Differential Revision: https://phorge.dev.yugabyte.com/D44574